requirements.txt
Flask==3.0.0
azure-cosmos==4.5.1
python-dotenv==1.0.0
gunicorn==21.2.0
openai==1.55.0
httpx==0.27.2
